requirements.txt
Flask-SQLAlchemy==2.4.1
SQLAlchemy==1.3.15
PyYAML==5.3.1
Flask==1.1.1
PyMySQL==0.9.3
Pandas==1.0.3
boto3==1.13.7
s3fs==0.4.2  
scikit-learn==0.22.1
numpy==1.16.3
pytest==5.4.1 
nltk==3.4.5
num2words==0.5.2
matplotlib==3.2.1
seaborn==0.10.1


